Fix rare bug in the F-based strategies of the pan module
In the pan module, strategies that rely on taking representative sequences
have a problem when all sequences have the same representative (they map to
the same seed). This was detected by checking if the first row of the seed
matrix consisted of all 1's. But, in very rare cases, this detection method
failed. This fix assesses directly if all sequences map to the same seed and
if yes, the family is not split.