Fix bug in pan module related to dna vs protein sequences
By default, MAFFT uses the first few characters of an alignment to decide
whether it's a protein or dna sequence. However, a protein sequence with
exclusively G, C, T and/or A amino acids in the beginning is wrongly
interpreted as a dna sequence. As a consequence, other amino acid characters
further in the sequence give an error. This was fixed for the pan module by
just telling MAFFT that it should assume protein sequences.