Fix occasional non-split of high-copy families
The splitting criterion of the pan module compares the expected genome overlap
(EGO) with the observed genome overlap (OGO) for the two subfamilies. The EGO
was calculated in such a way that it could occasionally be larger than one for
very high-copy gene families. This lead to a non-split, by definition, while
the OGO for such families is mostly one. The problem was fixed by reducing EGO
values greater than one to one.